Gemini/app2/requirements.txt
streamlit
google-generativeai
python-dotenv
youtube_transcript_api